requirements.txt
#requirements.txt
numpy==1.21.6
tensorflow-cpu==2.11.0